// src/web-app/App.types.ts
export enum AppBootstrapPhase {
  INITIAL = "initial",
  CONNECTING = "connecting",
  CONNECTED = "connected",
  STARTING = "starting",
  STARTED = "started",
  STOPPING = "stopping",
  STOPPED = "stopped",
  READY = "ready",
  FAILED = "failed",
}

export enum AppTheme {
  DARK = "bp6-dark",
  LIGHT = "bp6-light",
}